move calendar files to package
