+++ backend/src/main/java/org/example/backend/calendar/CalendarController.java
@@ -1,4 +1,4 @@
-package org.example.backend;
+package org.example.backend.calendar;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.GetMapping;
